Try to use bsdtar included with most Windows 10s before falling back to 7zip.
cache.h: Add missing constraint to get_lazy and change to use std::less<> for transparency. (Allows ToolCache to take StringView)
expected.h/expected.h: Teach Expected about LocalizedString
versions.h/versions.cpp: Teach versions parsing about LocalizedString and add a DotVersion parsing function that accepts version numbers like 19.00 (that 7zip uses)
util.h: Add a stream insertion operator for LocalizedString so that it can be used in INFO assertions in catch2.
tools.cpp:
* Localize
* Change to reuse existing version parsing machinery
* Change to not fail fast when a tool can't be found
* Fixed bug where git versions were being compared incorrectly; vcpkgTools.xml wants 2.17.1.windows.2 to be 2.17.1.2 but we only compared the first 3 numbers before for the minimum version check.
archive.cpp:
* Use the system's tar if available for zip files rather than downloading 7z for that
* Use CMake's tar if available for 7z files rather than downloading 7z for that
The last user of 7z is now PortableGit.

+++ src/vcpkg/commands.zbootstrap-standalone.cpp
@@ -47,9 +47,16 @@ namespace vcpkg::Commands
 #endif // ^^^ !VCPKG_STANDALONE_BUNDLE_SHA
 
         auto tool_cache = get_tool_cache(RequireExactVersions::NO);
-        const auto tar = tool_cache->get_tool_path_from_system(fs, Tools::TAR);
-        extract_tar(tar, bundle_tarball, vcpkg_root);
-        fs.remove(bundle_tarball, VCPKG_LINE_INFO);
-        Checks::exit_success(VCPKG_LINE_INFO);
+        const auto maybeTar = tool_cache->get_tool_path_from_system(fs, Tools::TAR);
+        if (maybeTar.has_value())
+        {
+            extract_tar(maybeTar.value_or_exit(VCPKG_LINE_INFO), bundle_tarball, vcpkg_root);
+            fs.remove(bundle_tarball, VCPKG_LINE_INFO);
+            Checks::exit_success(VCPKG_LINE_INFO);
+        }
+        else
+        {
+            Checks::exit_with_message(VCPKG_LINE_INFO, maybeTar.error());
+        }
     }
 }
